pyproject.toml
[tool.poetry]
name = "shiokobu"
version = "0.1.0"
description = "A simple wrapper to allow to pickle of MeCab object."
authors = ["nagomiso <daiseigi.giants+github@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
natto-py = "^0.9.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
pylint = "^2.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
